fix(security): per-request nonce CSP via middleware (#16)
Production script-src included 'unsafe-inline' unconditionally, defeating
CSP's protection against injected inline scripts. A per-request nonce
requires middleware, since static next.config headers can't mint one.

- add src/middleware.ts: mint a per-request nonce, emit a CSP with
  script-src 'self' 'nonce-...' (no 'unsafe-inline'); 'unsafe-eval'
  stays dev-gated, style-src keeps 'unsafe-inline'
- move CSP out of next.config.ts (keep the other static security
  headers there)

PostHog and Stripe load at runtime via their JS SDKs (no inline scripts)
and are origin-allowlisted, so no nonce attribute was needed. Verified
via check-types + build; recommend a browser CSP-console smoke check
before shipping.
