+++ apps/web/next.config.ts
@@ -1,35 +1,8 @@
 import type { NextConfig } from "next";
 
-const isDev = process.env.NODE_ENV === "development";
-
-// 'unsafe-eval' is dev-only (Next.js HMR / React refresh). Production bundles do not require it.
-// us-assets.i.posthog.com serves the PostHog array-loader assets; us.i.posthog.com is ingest.
-// worker-src 'self' blob: is required for PostHog session-recording web workers.
-const scriptSrc = [
-  "'self'",
-  "'unsafe-inline'",
-  isDev && "'unsafe-eval'",
-  "https://js.stripe.com",
-  "https://connect-js.stripe.com",
-  "https://*.posthog.com",
-  "https://us-assets.i.posthog.com",
-]
-  .filter(Boolean)
-  .join(" ");
-
-const csp = [
-  "default-src 'self'",
-  `script-src ${scriptSrc}`,
-  "connect-src 'self' https://api.stripe.com https://*.posthog.com https://us-assets.i.posthog.com https://api.resend.com https://utfs.io https://*.uploadthing.com",
-  "worker-src 'self' blob:",
-  "frame-src 'self' https://js.stripe.com https://hooks.stripe.com",
-  "img-src 'self' data: blob: https:",
-  "style-src 'self' 'unsafe-inline' https://fonts.googleapis.com",
-  "font-src 'self' https://fonts.gstatic.com",
-  "base-uri 'self'",
-  "form-action 'self'",
-].join("; ");
-
+// Content-Security-Policy is set per-request in `src/middleware.ts` (it mints a
+// per-request script-src nonce, which a static header here cannot do). The
+// remaining static security headers stay below.
 const securityHeaders = [
   {
     key: "Strict-Transport-Security",
@@ -46,10 +19,6 @@ const securityHeaders = [
   {
     key: "X-Frame-Options",
     value: "DENY",
-  },
-  {
-    key: "Content-Security-Policy",
-    value: csp,
   },
 ];
